DO NOT MERGE: Track pod resources according to lifecyle and node
The resource model of Kubernetes is sufficiently complex that it
can be hard to leverage the simple metric presented by spec to
make meaningful queries that reflect running behavior.

Expose a new pod_resource_requests metric that reports each pods
requests and limits in terms of the effective phase of the lifecyle
the pod is undergoing. This reports two series per resource - one
for the current phase (how much is seen by the component that owns
that resource usage) and one for the effective max as calculated
in use by the system. There are four key phases, pending (unscheduled)
where the scheduler has to make a decision, then initializing where
the pod is starting up, and then running where all containers are
considered, then terminal, where no resources are consumed at all.

This is still under active experimentation.